% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorrelations.R
\name{autocor_max_lag_table}
\alias{autocor_max_lag_table}
\title{autocor_max_lag_table}
\usage{
autocor_max_lag_table(
  data_file,
  data_file_2 = NULL,
  number_of_plots,
  need_abbr = FALSE,
  state = NULL,
  maxLag = 1,
  mean_threshold = 10,
  pVal = 0.05
)
}
\arguments{
\item{data_file}{Name of .csv file with data. No defaults.}

\item{number_of_plots}{Number of 1square meter plots in sampled area. No defaults}

\item{need_abbr}{Boolean. Whether species names should be abbreviated. Default to FALSE.}

\item{state}{Character. "g" - generative, "v" - vegetative, "v+j" - vegetative and juvenile.
Which states should be selected. If NULL, all shoots will be selected.
Dafault to NULL.}

\item{mean_threshold}{Number. Mean shoot number threshhold
for species to be taken into consideration. Default to 3.2.}

\item{pVal}{Number (0.05, 0.01 or less). p-value of Ljung-Box test
for species to be taken into consideration. Default to 0.05.}
}
\value{
Data frame with autocorrelation coefficients for maximum lag value,
where lag > 1.
}
\description{
autocor_max_lag_table
}
