% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorrelations.R
\name{autocorrelations}
\alias{autocorrelations}
\title{autocorrelations}
\usage{
autocorrelations(
  data_file,
  what_autocorr = "max_lag",
  number_of_plots,
  need_abbr = FALSE,
  state = NULL,
  mean_threshold = 3.2,
  pVal = 0.05
)
}
\arguments{
\item{data_file}{Name of .csv file with data. No defaults.}

\item{what_autocorr}{What lag we need. "max_lag" or "order_1". Default to "max_lag".}

\item{number_of_plots}{Number of 1square meter plots in sampled area. No defaults}

\item{need_abbr}{Boolean. Whether species names should be abbreviated. Default to FALSE.}

\item{state}{Character. "g" - generative, "v" - vegetative, "v+j" - vegetative and juvenile.
Which states should be selected. If NULL, all shoots will be selected.
Dafault to NULL.}

\item{mean_threshold}{Number. Mean shoot number threshhold
for species to be taken into consideration. Default to 3.2.}

\item{pVal}{Number (0.05, 0.01 or less). p-value of Ljung-Box test
for species to be taken into consideration. Default to 0.05.}
}
\value{
Data frame with autocorrelation coefficients for maximum lag value.
}
\description{
autocorrelations
}
