% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_where.R
\name{check_where}
\alias{check_where}
\title{check_where()}
\usage{
check_where(data_file, sp_name, sample_pattern = NULL)
}
\arguments{
\item{data_file}{A name of .csv file with the data. No defaults.}

\item{sp_name}{A full name of the species as it is provided in the data file. No defaults.}

\item{sample_pattern}{A pattern for a certain sample name. Default to NULL.}
}
\value{
table for species occurens at the samples or data.frame for the species occurence at a certain sample
}
\description{
This function is for check on what sample certain species present and with what frequency.
Also, it can provide when the species was found at the certain sample
}
\examples{
check_where(data_file = "foo.csv", sp_name = ".Leontodon hispidus L.")
# returns frequency table for presence at samples

check_where(data_file = "foo.csv", sp_name = ".Leontodon hispidus L.", sample_pattern = "^I_2_")
# returns data.frame with the species occurence at a certain sample

teberda.tools::check_where()

}
