% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tidy_data.R
\name{get_abbr}
\alias{get_abbr}
\title{This function allows to get abbreviations of species' names}
\usage{
get_abbr(sp_name, start1 = 1, end1 = 3, start2 = 1, end2 = 7)
}
\arguments{
\item{sp_name}{A vector of full latin species names with authors. No defaults.}

\item{start1}{A number for starting symbol in the abbriviation of the genus. Defaults to 1.}

\item{end1}{A number for ending symbol in the abbriviation of the first word. Default to 4.}

\item{start2}{A number for starting symbol in the abbriviation of the second word. Default to 2.}

\item{end2}{A number for ending symbol in the abbriviation of the second word. Default to 7.}
}
\description{
This function allows to get abbreviations of species' names
}
