% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tidy_data.R
\name{get_long_table}
\alias{get_long_table}
\title{Gets wide table with shoot numbers of every state on every sample
through the years of observation. Returns a tidy long variant of this table}
\usage{
get_long_table(wide_table_df)
}
\arguments{
\item{wide_table_df}{A wide table got from raw .csv.
get_long_table()}
}
\description{
Gets wide table with shoot numbers of every state on every sample
through the years of observation. Returns a tidy long variant of this table
}
