% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tidy_data.R
\name{get_wide_transponed}
\alias{get_wide_transponed}
\title{Takes long_data
Returns wide_transponed_data (column names are species, row names are years)}
\usage{
get_wide_transponed(long_data)
}
\arguments{
\item{long_data}{The result of get_long_table() function.
get_wide_transponed()}
}
\description{
Takes long_data
Returns wide_transponed_data (column names are species, row names are years)
}
