% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tidy_data.R
\name{get_without_zeros}
\alias{get_without_zeros}
\title{Gets rid of zeros in sumed data.
Takes transponed data frame, returns the data frame with the only columns
which means are more than 0 (without_zeros)}
\usage{
get_without_zeros(transponed_df)
}
\arguments{
\item{transponed_df}{The result of get_wide_transponed().}
}
\description{
Gets rid of zeros in sumed data.
Takes transponed data frame, returns the data frame with the only columns
which means are more than 0 (without_zeros)
}
