% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_scatt_graph.R
\name{graph_all_shoot_num}
\alias{graph_all_shoot_num}
\title{Graph the numbers of shoots vs years}
\usage{
graph_all_shoot_num(
  sp_name_list,
  csv_high,
  plot_title,
  csv_low = NULL,
  state = NULL,
  tr_line = lm,
  mute_ax_lab = TRUE,
  font_face = "bold",
  font_family = "Arial",
  main_font_size = 16,
  axis_font_size = 14,
  ...
)
}
\arguments{
\item{sp_name_list}{A character vector contanes latin names of species in question. No default.}

\item{csv_high}{Name of .csv file with data. No defaults.}

\item{plot_title}{The main title of the plot. No default.}

\item{csv_low}{Name of second .csv file. Default to NULL}

\item{state}{Character. "g" - generative, "v" - vegetative, "v+j" - vegetative and juvenile.
Which states should be selected. If NULL, all shoots will be selected.
Dafault to NULL.}

\item{tr_line}{What trend line method should be applied to the charts. Default to lm.}

\item{mute_ax_lab}{Boolean. Whether we should mute axis names to each indivdual scatterplot.
Default to TRUE.}

\item{font_face}{Default to "bold".}

\item{font_family}{Default to "Arial".}

\item{main_font_size}{Default to 16.}

\item{axis_font_size}{Default to 14.}
}
\value{
arranged scatterplots
}
\description{
Combines scatterplots for all species in sp_list into one image
with common main heading and common axis names
}
