% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltrend.R
\name{ltrend}
\alias{ltrend}
\title{ltrend}
\usage{
ltrend(
  data_file,
  number_of_plots,
  need_abbr = FALSE,
  state = NULL,
  threshold_mean = 3.2,
  pVal = 0.05
)
}
\arguments{
\item{data_file}{Name of .csv file with data. No defaults.}

\item{number_of_plots}{Number of 1square meter plots in sampled area. No defaults}

\item{need_abbr}{Boolean. Whether species names should be abbreviated. Default to FALSE.}

\item{state}{Character. "g" - generative, "v" - vegetative, "v+j" - vegetative and juvenile.
Which states should be selected. If NULL, all shoots will be selected.
Dafault to NULL.}

\item{pVal}{Number (0.05, 0.01 or less). p-value of anovaANOVA comparing two gls models compared with ML methods
for species to be taken into consideration. Default to 0.05.}

\item{mean_threshold}{Number. Mean shoot number threshhold
for species to be taken into consideration. Default to 3.2.}
}
\value{
dataframe with linear trends parameters
}
\description{
ltrend
}
