% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_scatt_graph.R
\name{shoots_as_line}
\alias{shoots_as_line}
\title{Lines for numbers of shoots}
\usage{
shoots_as_line(
  csv_high,
  csv_low = NULL,
  sp_names,
  state = NULL,
  plot_title = NULL,
  mute_ax_lab = FALSE,
  loged = FALSE,
  ...
)
}
\arguments{
\item{csv_high}{Name of .csv file with data. No defaults.}

\item{csv_low}{Name of second .csv file. Default to NULL}

\item{sp_names}{A character vector with the names of the species of interest}

\item{state}{Character. "g" - generative, "v" - vegetative, "v+j" - vegetative and juvenile.
Which states should be selected. If NULL, all shoots will be selected.
Dafault to NULL.}

\item{plot_title}{The main title of the plot. No default.}

\item{mute_ax_lab}{Boolean. Whether we should mute axis names to each indivdual scatterplot.
Default to TRUE.}

\item{loged}{Boolean. Whether the logarithm of number of shoots should be taken. Defaul to FALSE.}
}
\value{
ggplot2 object with line
}
\description{
Plot a line for the number of shoots in the years of observation.
}
