% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_scatt_graph.R
\name{shoots_scatter_plot}
\alias{shoots_scatter_plot}
\title{Scatterplot for a species}
\usage{
shoots_scatter_plot(
  csv_high,
  csv_low = NULL,
  sp_name,
  state = NULL,
  tr_line = lm,
  plot_title = sp_name,
  mute_ax_lab = TRUE,
  ...
)
}
\arguments{
\item{csv_high}{Name of .csv file with data. No defaults.}

\item{csv_low}{Name of second .csv file. Default to NULL}

\item{state}{Character. "g" - generative, "v" - vegetative, "v+j" - vegetative and juvenile.
Which states should be selected. If NULL, all shoots will be selected.
Dafault to NULL.}

\item{tr_line}{What trend line method should be applied to the charts. Default to lm.}

\item{plot_title}{The main title of the plot. No default.}

\item{mute_ax_lab}{Boolean. Whether we should mute axis names to each indivdual scatterplot.
Default to TRUE.}
}
\value{
Scatterplot with the trend line
}
\description{
Plot scatter diagrams of the number of shoots in the years of observation.
}
