% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speccorr.R
\name{speccorr}
\alias{speccorr}
\title{speccorr}
\usage{
speccorr(
  data_file,
  data_fie2 = NULL,
  need_abbr = FALSE,
  state = NULL,
  type = "pearson",
  pVal = 0.05
)
}
\arguments{
\item{data_file}{Name of .csv file with data. No defaults.}

\item{need_abbr}{Boolean. Whether species names should be abbreviated. Default to FALSE.}

\item{state}{Character. "g" - generative, "v" - vegetative, "v+j" - vegetative and juvenile.
Which states should be selected. If NULL, all shoots will be selected.
Dafault to NULL.
@param type Type of correlation coefficient ("pearson" or "spearman").
Default to "pearson".}

\item{pVal}{Number (0.05, 0.01 or less). p.value of corr.test
for species to be taken into consideration. Default to 0.05.}

\item{data_file_2}{Name of second .csv file to be united with the first one.
It must be checked, that the data are from the SAME period.
Default to NULL}
}
\value{
Data frame with correlation coefficients and their p.values between species
}
\description{
Correlations between species.
Helper function `flat_cor_mat` comes from
https://rstudio-pubs-static.s3.amazonaws.com/240657_5157ff98e8204c358b2118fa69162e18.html
Rename columns added.
}
