% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tidy_data.R
\name{states_selector}
\alias{states_selector}
\title{Function for selection the state of shoots if any.
Checks what state(s) should be selected and makes data frame for it (or them together)}
\usage{
states_selector(long_table, state = NULL)
}
\arguments{
\item{long_table}{From get_long_table().}

\item{state}{"g" - generative, "v" - vegetative, "v+j" - vegetative and juvenile. Default to NULL.
states_selector()}
}
\description{
Function for selection the state of shoots if any.
Checks what state(s) should be selected and makes data frame for it (or them together)
}
