% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vggvprinextcorr.R
\name{vggvprinextcorr}
\alias{vggvprinextcorr}
\title{vggvprinextcorr}
\usage{
vggvprinextcorr(
  data_file,
  data_file_2 = NULL,
  comparing_mod = "v_vs_g",
  method = "spearman",
  pVal = 0.05
)
}
\arguments{
\item{data_file}{Name of .csv file with data. No defaults.}

\item{data_file_2}{Name of second .csv file to be united with the first one.
It must be checked, that the data are from the SAME period.
Default to NULL}

\item{comparing_mod}{Key signs which modification of test should be done.
It can be:
  "g_vs_pyv" number of generative shoots with prior-year number of vegetative shoots
  "v_vs_pyg" number of vegetative shoots with prior-year number of generative shoots
  "v_vs_g"   number of vegetative shoots with number of generative shoots
   of the same year.
   Default to "v_vs_g".}

\item{method}{The method of estimation of correlation coeffitients.
Can be "spearman", "pearson" or "kendall". Default to "spearman".}

\item{pVal}{Number (0.05, 0.01 or less). p.value of corr.test
for species to be taken into consideration. Default to 0.05.}
}
\value{
Data frame with correlation coefficients.
}
\description{
The function estimates correlations between number of vegetative shoots
in the prior year and generative shoots of the next year and vise versa.
Also the correlations can be estimated for generative and vegetative shoot numbers
of the same year (default behavior).
}
